Plugins: Allow plugins to add an interface message
This adds the function delegate AddMessage to the openBVE API, and moves the definition of MessageColor to OpenBveApi.Colors
Must be called as follows:
AddMessage(Message,Color,Time)
Message should be a string representing the message you wish displayed.
Color should be the color you wish the message to display in.
Time should be the time in seconds for your message to display.