Revert a paralellization change.
This seems to cause intermittent issues with BVEC_ATS's debug window. (Control created on different thread exceptions)
Whilst this window is buggy in the first place, there are various trains out of my control using it.....

NOTE: 
A plugin overlay similar to the Route Information overlay is on my 'list' of things to look at, which would deprecate this window.