Packages:
Fix: Don't crash if we click on the 'Replace Package' button with nothing selected whilst creating a package.
Add prompt that there are no installed packages if the database is empty, and the user asks to replace an existing package.
Add prompt querying if the user would like to remove the package from the database if the uninstall fails.