Fix: Reordering the columns on a package list now works.
Fix: We should clone the object rather than passing the reference when adding it to a dependancy list.
Change: Dependancy lists don't include the package version number.
Change: If the BackgroundWorker thread is still busy, sleep whilst we wait for it to terminate. (Attempt to fix some odd errors, but if we're unlucky this will cause more trouble.....)