Catch error on bug report dialog
Apparently there was always a follow up message (Even if not as detailed)

But only shows if you have a PackageDatabase folder as DirectoryInfo.GetFiles throws an DirectoryNotFound exception, which then is silently ignored.

This commit fixes it so it will check if the directory exist, and will also result in a prompt if failed.