Fix: Applying the brakes at very slow speeds on level track didn't work
The intent of the original code appears to be to clamp the max deceleration to that of gravity when below 0.01km/h (presumably to simulate a smooth stop), but that failed to take into account the case where brakes were applied at very low speeds (~0.05km/h and below).
In this case, our original deceleration might take us to just below the 0.01km/h figure, and the brakes are no longer effective, leaving the train fractionally moving.